Fix unit tests (#12)
