+++ test/test_extensions.rb
@@ -5,52 +5,10 @@
 
 require 'asciidoctor'
 require 'test/unit'
-require_relative '../../stages/test/main_module.rb'
-require_relative '../../stages/test/cli.rb'
-Dir['../../stages/test/modules.d/*.rb'].each { |file| require file }
-
-def assert_any_startwith(errors, text)
-  assert_true(errors.any? { |err| err[:msg].start_with?(text) })
-end
-
-def init(content, name)
-  document, _original = init2(content, name)
-  return document
-end
-
-def init2(content, name, filename = nil)
-  filename = name + '.adoc' if filename.nil?
-  tempfile_path = write_tempfile(filename, content)
-  document = Asciidoctor.load_file(tempfile_path, safe: :safe, catalog_assets: true)
-  document.convert
-  original = Asciidoctor.load_file(tempfile_path, safe: :safe, catalog_assets: true)
-  return document, original
-end
-
-class TestPatternBlacklist < Test::Unit::TestCase
-  def test_pattern_blacklist
-    adoc = '= Bad lines
-
-======= too long heading
-WPP
-document-center
-
-    '
-    blacklist_patterns = '
-# do not match this comment
-// do not match this comment
-/document-center/
-/WPP/
-document
-/={6,}/
-/bad_word/
-    '
-    blacklist_file_path = write_tempfile('blacklist_patterns.txt', blacklist_patterns)
-    document, original = init2(adoc, "#{self.class.name}_#{__method__}", 'test_toolchain_pattern_blacklist.adoc')
-    errors = Toolchain::PatternBlacklist.new.run(document, original, blacklist_file_path)
-    assert_equal(3, errors.length)
-  end
-end
+require_relative '../lib/stages/test.rb'
+require_relative './util.rb'
+extensions_dir = File.join(__dir__, '..', 'lib', 'extensions.d', '*.rb')
+Dir[extensions_dir].each { |file| require file }
 
 class TestIDChecker < Test::Unit::TestCase
   def test_short_ids
@@ -110,5 +68,50 @@ Thank you.
       endc = msg.index("'", startc) - 1
       msg[startc..endc]
     end
+  end
+end
+
+class TestLinkChecker < Test::Unit::TestCase
+  def test_links
+    omit_if(ENV.key?('SKIP_NETWORK'), 'Tests with networking disabled')
+    adoc = '= Test links
+
+1. https://github.com/wirecard/docs-toolchain[Docs Toolchain]
+2. https://github.com/asciidoctor/asciidoctor-exteansions-lab[Asciidoctor Extensions Lab]
+3. https://adfasdgea.asd/adfadfasdf/[Unknown Domain]
+4. http://111.222.123.48[Random IP]
+    '
+    document = init(adoc, self.class.name)
+    assert_equal(4, document.references[:links].length)
+    errors = Toolchain::LinkChecker.new.run(document)
+    assert_equal(3, errors.length)
+    assert_any_startwith(errors, '[404] Not Found') # 2.
+    assert_any_startwith(errors, 'SocketError') # 3.
+    assert_any_startwith(errors, 'Net::OpenTimeout') # 4.
+  end
+end
+
+class TestPatternBlacklist < Test::Unit::TestCase
+  def test_pattern_blacklist
+    adoc = '= Bad lines
+
+======= too long heading
+WPP
+document-center
+
+    '
+    blacklist_patterns = '
+# do not match this comment
+// do not match this comment
+/document-center/
+/WPP/
+document
+/={6,}/
+/bad_word/
+    '
+    blacklist_file_path = write_tempfile('blacklist_patterns.txt', blacklist_patterns)
+    document, original = init2(adoc, "#{self.class.name}_#{__method__}", 'test_toolchain_pattern_blacklist.adoc')
+    errors = Toolchain::PatternBlacklist.new.run(document, original, blacklist_file_path)
+    assert_equal(3, errors.length)
   end
 end
